Update bdqcore_quickreference-header.md
Rewrite to simplify to what is important for users wanting to know the guts of BDQ Core. That is align with the simplicity and focus of Darwin Core. Include links to other key documents: Test terms/definitions, Introduction, users' and implementers' guides and vocabularies.

I would GREATLY prefer to remove all bar the Title in the Header- 'Date version issued' as there is a ref in the text, 'Part of the TDWG standard', Abstract (we REALLY don't need this!), Contributors, Creator (use Citation footer) and Comment. This will be my next edit!