Update users-header.md
Complete review. Some duplication removed and some rewriting (e.g., passive to active voice) for clarity.

Do we need Section 5, given the comprehensive discussion of this and related issues in the Supplement?

I can't see anything major omitted, but the discrimination about the content of this, Implementers and Supplement should be examined.