Update README.md
Added a placeholder for 'Notes for the Review manager' so we don't forget it. (But as soon as we know who this will be, I will send a link to the notes). This is also a reminder to add how the Review Manager wants to deal with the feedback process. This will need to be added to the bottom of this document.